Share shutdown result predicates
